♻️ RC1DX8 refactor: split Hermes command catalog
