+++ .agentplane/tasks/202607221908-RC1DX8/pr/meta.json
@@ -2,7 +2,7 @@
   "base": "main",
   "branch": "task/202607221908-RC1DX8/migrate-runner-and-hermes-command-boundaries",
   "created_at": "2026-08-01T02:58:47.934Z",
-  "diffstat_sha256": "sha256:4eeb7c8d26ecc0bb88024010c659fcb770498d72dc9808ccee5aa6255ce890d1",
+  "diffstat_sha256": "sha256:e822e0bb92715b9e21c382fa6ce0d8d3177937c02d53f793682a1aab67ff838a",
   "last_verified_at": "2026-08-01T11:12:25.369Z",
   "last_verified_diffstat_sha256": "sha256:4eeb7c8d26ecc0bb88024010c659fcb770498d72dc9808ccee5aa6255ce890d1",
   "pr_number": 4717,
